requirements.txt
langchain_google_genai
langchain_core
langchain_community
python-dotenv
PyPDF2